Fix #77728: pgort140.DLL missing
If a PGO build is requested, but a build error occurs after the PGI
build has succeeded and before the PGO build has even been attempted
(for instance, if the PGO training fails), we may get a bogus archive
containing the PGI build, without any hint.  We apply a quick fix which
requires to run the PGO build if requested, and reports a respective
error otherwise.