Add general support for exporting from a Github mirror
This is a follow up on the quick fix for exporting from the php-src
Github mirror[1].  Instead of hard-coding everything in the code, we
introduce a configurable `gh_url`, and use it if given.  We cannot use
the repository URL, though, because the toplevel folder inside the
downloaded archive would have a different name (the shortened revision
number is expanded again); downloading from the codeload subdomain
gives the desired result.

[1] <https://github.com/php/web-rmtools/commit/9e1ca1002ade165431925eca3c6ec898244de02c>