Prepare for memcached builds
In theory, this should work, but in practice it fails since libmemcached
ships headers with standard names[1] which are erroneously included.  We
still make the change, if only as a reminder that libmemcached needs to
be fixed.

We shall have a look at building with `--enable-memcached-igbinary` when
the header issue has been resolved.

[1] <https://github.com/awesomized/libmemcached/issues/61#issuecomment-1063228187>